// src/frontend-applications/agent-portal/modules/webform/model/WebformProperty.ts
/**
 * Copyright (C) 2006-2022 c.a.p.e. IT GmbH, https://www.cape-it.de
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum WebformProperty {

    ID = 'ObjectId',

    BUTTON_LABEL = 'buttonLabel',

    TITLE = 'title',

    SHOW_TITLE = 'showTitle',

    SAVE_LABEL = 'saveLabel',

    HINT_MESSAGE = 'hintMessage',

    SUCCESS_MESSAGE = 'successMessage',

    MODAL = 'modal',

    USE_KIX_CSS = 'useKIXCSS',

    ALLOW_ATTACHMENTS = 'allowAttachments',

    ACCEPTED_DOMAINS = 'acceptedDomains',

    QUEUE_ID = 'QueueID',

    PRIORITY_ID = 'PrioritiyID',

    TYPE_ID = 'TypeID',

    STATE_ID = 'StateID',

    USER_LOGIN = 'userLogin',

    USER_PASSWORD = 'webformUserPassword'

}